Extracted frame_options to separate class
